% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FA_process}
\alias{FA_process}
\title{Data preprocessing}
\usage{
FA_process(
  data,
  scaled = FALSE,
  coding = NA,
  pareto_scale = FALSE,
  log_trans = FALSE
)
}
\arguments{
\item{data}{The data to be processed. If not a matrix, will be transformed in to a matrix.}

\item{scaled}{TRUE/FALSE. Should the input data be scaled? Defaults to FALSE.}

\item{coding}{The coding used to indicate a data entry is missing. Defaults to NA.}

\item{pareto_scale}{TRUE/FALSE. Should the input data be Pareto scaled? Defaults to FALSE.}

\item{log_trans}{TRUE/FALSE. Should the input data be log transformed? Defaults to FALSE.}
}
\value{
A list containing three entries.

* "data" contains the processed data in matrix format with missing entries coded as NA.

* "missingness_pattern" contains the missingness pattern of the data for use in the tIFA model.

* "divisors" contains a vector of the divisors used to scale the input data, if applicable. If unscaled, the value will be NULL.
}
\description{
Processes a dataset for use in a tIFA model.
}
\examples{
example_data <- matrix(rnorm(20), nrow = 5)
example_data[4, 2] <- 0.001  # add missingness to example data coded as 0.001
FA_process(example_data, coding = 0.001)
}
