% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The mode of x. If x is multi-modal, the first mode (by index in x) is returned.
}
\description{
Mode
}
\examples{
x <- c(1, 2, 3, 4, 4, 5)
Mode(x)
}
