% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tifa_post.R
\name{tifa_post}
\alias{tifa_post}
\title{Post-processing for tIFA results}
\usage{
tifa_post(tifa_res, burn)
}
\arguments{
\item{tifa_res}{The output object of the [tIFA_model()] function.}

\item{burn}{The desired burn-in.}
}
\value{
Returns a list containing two entries.

* imputed_dataset: The returned dataset in matrix format with the imputed values inserted.

* imputation_info: A dataframe containing information on imputed values and associated uncertainty for each missing entry in the original dataset.
}
\description{
Post-processing for tIFA results
}
\examples{
# generate example data
example_data <- matrix(abs(rnorm(100)), nrow = 5)

# add missingness to example data coded as 0.001
example_data[4, 2] <- 0.001
example_data[2, 18] <- 0.001

# apply data pre-processing function
input_data <- FA_process(example_data, coding = 0.001)

# perform zero imputation to use as initial imputed data value in tIFA
zero_imp <- input_data$data
zero_imp[is.na(zero_imp)] <- 0

# run tIFA model (short chain for example)
tifa_res <- tIFA_model(data = input_data$data, missingness = input_data$missingness, M = 1000,
                       initial_dataset = zero_imp, k = 5)

# process tIFA results (short burn for example)
processed_res <- tifa_post(tifa_res, burn = 10)
}
